fix tune_cfg issue for 3.x static quant
