+++ neural_compressor/torch/algorithms/static_quant/utility.py
@@ -16,8 +16,11 @@ import copy
 import json
 import os
 import re
+import subprocess
 from typing import Dict, List, Union
 
+import cpuinfo
+import psutil
 import torch
 from packaging.version import Version
 
@@ -63,57 +66,238 @@ BLOCK_PATTERNS = [
 ]
 
 
-def cfg_to_qconfig(tune_cfg, cfgs, default_cfgs, fuse_ops):  # pragma: no cover
+def cfg_to_qconfig(
+    tune_cfg, cfgs, default_cfgs, fuse_ops, op_infos_from_cfgs, output_tensor_id_op_name
+):  # pragma: no cover
     assert cfgs is not None, "No configure for IPEX int8 model..."
-    for key in tune_cfg["op"]:
-        try:
-            scheme = tune_cfg["op"][key]["activation"]["scheme"]
-        except:
-            scheme = "asym"
-        if scheme not in ["asym", "sym"]:
-            scheme = "asym"
-        break
-    for key in tune_cfg["op"]:
-        value = tune_cfg["op"][key]
-        pattern = get_pattern(key, fuse_ops)
-        assert isinstance(value, dict)
-        assert "activation" in value
-        if value["activation"]["dtype"] == "fp32":
-            if "weight" in value:
-                assert value["weight"]["dtype"] == "fp32"
-            for op_cfg in cfgs:
-                if op_cfg["id"] == key[0]:
-                    if key[1] in ["relu_", "add_"]:
-                        continue
-                    num_inputs = len(op_cfg["inputs_quantized"])
-                    num_outputs = len(op_cfg["outputs_quantized"])
-                    for i_num in range(num_inputs):
-                        op_cfg["inputs_quantized"][i_num] = False
-                    for o_num in range(num_outputs):
-                        op_cfg["outputs_quantized"][o_num] = False
-                    if pattern:
-                        if pattern[1] in ["relu_", "add_"]:
+    if ipex_ver.release < Version("1.12.0").release:  # pragma: no cover
+        for key in tune_cfg["op"]:
+            try:
+                scheme = tune_cfg["op"][key]["activation"]["scheme"]
+            except:
+                scheme = "asym"
+            if scheme not in ["asym", "sym"]:
+                scheme = "asym"
+            break
+        for key in tune_cfg["op"]:
+            value = tune_cfg["op"][key]
+            pattern = get_pattern(key, fuse_ops)
+            assert isinstance(value, dict)
+            assert "activation" in value
+            if value["activation"]["dtype"] == "fp32":
+                if "weight" in value:
+                    assert value["weight"]["dtype"] == "fp32"
+                for op_cfg in cfgs:
+                    if op_cfg["id"] == key[0]:
+                        if key[1] in ["relu_", "add_"]:
                             continue
-                        tune_cfg["op"][pattern]["activation"]["dtype"] = "fp32"
-                        if "weight" in tune_cfg["op"][pattern]:
-                            tune_cfg["op"][pattern]["weight"]["dtype"] = "fp32"
+                        num_inputs = len(op_cfg["inputs_quantized"])
+                        num_outputs = len(op_cfg["outputs_quantized"])
+                        for i_num in range(num_inputs):
+                            op_cfg["inputs_quantized"][i_num] = False
+                        for o_num in range(num_outputs):
+                            op_cfg["outputs_quantized"][o_num] = False
+                        if pattern:
+                            if pattern[1] in ["relu_", "add_"]:
+                                continue
+                            tune_cfg["op"][pattern]["activation"]["dtype"] = "fp32"
+                            if "weight" in tune_cfg["op"][pattern]:
+                                tune_cfg["op"][pattern]["weight"]["dtype"] = "fp32"
+            else:
+                for op_cfg in cfgs:
+                    if op_cfg["id"] == key[0]:
+                        if key[1] in ["relu_", "add_"]:
+                            continue
+                        num_inputs = len(op_cfg["inputs_quantized"])
+                        num_outputs = len(op_cfg["outputs_quantized"])
+                        for i_num in range(num_inputs):
+                            op_cfg["inputs_quantized"][i_num] = default_cfgs[key[0]]["inputs_quantized"][i_num]
+                        for o_num in range(num_outputs):
+                            op_cfg["outputs_quantized"][o_num] = default_cfgs[key[0]]["outputs_quantized"][o_num]
+        with open(ipex_config_path, "w") as write_f:
+            json.dump(cfgs, write_f)
+        if scheme == "asym":
+            return torch.per_tensor_affine
         else:
-            for op_cfg in cfgs:
-                if op_cfg["id"] == key[0]:
-                    if key[1] in ["relu_", "add_"]:
-                        continue
-                    num_inputs = len(op_cfg["inputs_quantized"])
-                    num_outputs = len(op_cfg["outputs_quantized"])
-                    for i_num in range(num_inputs):
-                        op_cfg["inputs_quantized"][i_num] = default_cfgs[key[0]]["inputs_quantized"][i_num]
-                    for o_num in range(num_outputs):
-                        op_cfg["outputs_quantized"][o_num] = default_cfgs[key[0]]["outputs_quantized"][o_num]
-    with open(ipex_config_path, "w") as write_f:
-        json.dump(cfgs, write_f)
-    if scheme == "asym":
-        return torch.per_tensor_affine
+            return torch.per_tensor_symmetric
+
     else:
-        return torch.per_tensor_symmetric
+        op_infos = copy.deepcopy(op_infos_from_cfgs)
+        cfgs = check_cfg_and_qconfig(tune_cfg["op"], cfgs, op_infos, output_tensor_id_op_name, smooth_quant=False)
+
+        with open(ipex_config_path, "w") as write_f:
+            json.dump(cfgs, write_f, indent=4)
+        return None
+
+
+def check_cfg_and_qconfig(
+    tune_cfg, cfgs, op_infos_from_cfgs, output_tensor_ids_op_name, smooth_quant=False
+):  # pragma: no cover
+    """Check configs and quantization configs.
+
+    Args:
+        tune_cfg (dict): dictionary of quantization configuration.
+        cfgs (dict): the input configs.
+        op_infos_from_cfgs (dict): op infos from configs.
+        output_tensor_ids_op_name (dict): dictionary of output tensor op names.
+
+    Returns:
+        cfgs (dict).
+    """
+    for op_name in tune_cfg:
+        inc_op_cfg = tune_cfg[op_name]
+        for i, name in enumerate(op_name[0]):
+            # to int8
+            ipex_op_cfg = op_infos_from_cfgs[name]
+            input_tensor_infos = ipex_op_cfg["input_tensor_infos"]
+            if op_name[1] == "Linear" or op_name[1] == "Linear&add":  # record op_name for possible op-wise fallback
+                logger.debug(f"ipex_op_cfg['fqn'] - op_name {ipex_op_cfg['fqn']}  {op_name}")
+            for index, input_tensor_info in enumerate(input_tensor_infos):
+                if "force_dtype" not in input_tensor_info.keys():
+                    continue
+                if (
+                    input_tensor_info["force_dtype"] == "torch.qint8"
+                    or input_tensor_info["force_dtype"] == "torch.quint8"
+                ):
+                    # int8 -> int8
+                    if inc_op_cfg["weight"]["dtype"] == "int8":
+                        inc_scheme = inc_op_cfg["activation"]["scheme"]
+                        inc_algorithm = inc_op_cfg["activation"]["algorithm"]
+                        ipex_op_cfg["input_tensor_infos"] = input_tensor_infos
+                        if (
+                            "op_type" in ipex_op_cfg
+                            and ipex_op_cfg["op_type"] == "<class 'torch.nn.modules.linear.Linear'>"
+                        ):
+                            smooth_quant_enable = True
+                        else:
+                            smooth_quant_enable = False
+                        activation_observer = generate_activation_observer(
+                            inc_scheme, inc_algorithm, smooth_quant, smooth_quant_enable
+                        )
+                        if not smooth_quant:
+                            if inc_scheme == "sym":
+                                input_tensor_infos[index]["force_dtype"] = "torch.qint8"
+                            if inc_scheme == "asym":
+                                input_tensor_infos[index]["force_dtype"] = "torch.quint8"
+                        ipex_op_cfg["activation_observer"] = activation_observer
+                    # int8 -> fp32
+                    else:
+                        input_tensor_infos[index]["force_dtype"] = "torch.float32"
+                    # modify pre_op output inf_dtype
+                    if i == 0:
+                        input_tensor_id = input_tensor_info["id"]
+                        input_tensor_dtype = input_tensor_info["force_dtype"]
+                        if input_tensor_id in output_tensor_ids_op_name.keys():
+                            pre_op_name = output_tensor_ids_op_name[input_tensor_id]
+                            pre_op_module = pre_op_name[0][0]
+                            pre_op_state = pre_op_name[0][1]
+                            pre_op_index = pre_op_name[0][2]
+                            pre_op_infos = cfgs[pre_op_module][pre_op_state][pre_op_index]
+                            pre_op_output_infos = pre_op_infos["output_tensor_infos"]
+                            for index, pre_op_output in enumerate(pre_op_output_infos):
+                                if pre_op_output["id"] == input_tensor_id:
+                                    pre_op_output_infos[index]["inf_dtype"] = input_tensor_dtype
+                                else:
+                                    pass
+                            pre_op_infos["output_tensor_infos"] = pre_op_output_infos
+                            cfgs[pre_op_module][pre_op_state][pre_op_index] = pre_op_infos
+                        else:
+                            pass
+            cfgs[name[0]][name[1]][name[2]] = ipex_op_cfg
+    return cfgs
+
+
+def generate_activation_observer(scheme, algorithm, smooth_quant=False, smooth_quant_enable=False):  # pragma: no cover
+    """This is a helper method to generate an activation observer.
+
+    Args:
+        scheme (str): Quantization scheme to be used.
+        algorithm (str): What algorithm for computing the quantization parameters based on.
+
+    Returns:
+        An observer.
+    """
+    kl_activation_observer = {
+        "name": "HistogramObserver",
+        "bins": 2048,
+        "upsample_rate": 128,
+        "dtype": "torch.quint8",
+        "qscheme": "torch.per_tensor_affine",
+        "reduce_range": False,
+        "quant_min": 0,
+        "quant_max": 255,
+    }
+    minmax_activation_observer = {
+        "name": "MinMaxObserver",
+        "dtype": "torch.quint8",
+        "qscheme": "torch.per_tensor_affine",
+        "reduce_range": False,
+        "quant_min": 0,
+        "quant_max": 255,
+    }
+    smoothquant_kl_activation_observer = {
+        "name": "SmoothQuantActivationObserver",
+        "smooth_quant_enabled": smooth_quant_enable,
+        "dtype": "torch.quint8",
+        "qscheme": "torch.per_tensor_affine",
+        "reduce_range": False,
+        "quant_min": 0,
+        "quant_max": 255,
+        "alpha": 0.5,
+        "act_observer": kl_activation_observer,
+        "act_ic_observer": {
+            "name": "PerChannelMinMaxObserver",
+            "ch_axis": -1,
+            "dtype": "torch.quint8",
+            "qscheme": "torch.per_channel_affine",
+            "reduce_range": False,
+            "quant_min": 0,
+            "quant_max": 255,
+        },
+    }
+    smoothquant_minmax_activation_observer = {
+        "name": "SmoothQuantActivationObserver",
+        "smooth_quant_enabled": smooth_quant_enable,
+        "dtype": "torch.quint8",
+        "qscheme": "torch.per_tensor_affine",
+        "reduce_range": False,
+        "quant_min": 0,
+        "quant_max": 255,
+        "alpha": 0.5,
+        "act_observer": minmax_activation_observer,
+        "act_ic_observer": {
+            "name": "PerChannelMinMaxObserver",
+            "ch_axis": -1,
+            "dtype": "torch.quint8",
+            "qscheme": "torch.per_channel_affine",
+            "reduce_range": False,
+            "quant_min": 0,
+            "quant_max": 255,
+        },
+    }
+    REDUCE_RANGE = False if CpuInfo().vnni else True
+    if REDUCE_RANGE:
+        minmax_activation_observer["reduce_range"] = REDUCE_RANGE
+        kl_activation_observer["reduce_range"] = REDUCE_RANGE
+    if scheme == "sym":
+        minmax_activation_observer["qscheme"] = "torch.per_tensor_symmetric"
+        minmax_activation_observer["dtype"] = "torch.qint8"
+        minmax_activation_observer["quant_min"] = -128
+        minmax_activation_observer["quant_max"] = 127
+        kl_activation_observer["qscheme"] = "torch.per_tensor_symmetric"
+        kl_activation_observer["dtype"] = "torch.qint8"
+        kl_activation_observer["quant_min"] = -128
+        kl_activation_observer["quant_max"] = 127
+    if smooth_quant and smooth_quant_enable:
+        if algorithm == "kl":
+            return smoothquant_kl_activation_observer
+        if algorithm == "minmax":
+            return smoothquant_minmax_activation_observer
+    else:
+        if algorithm == "kl":
+            return kl_activation_observer
+        if algorithm == "minmax":
+            return minmax_activation_observer
 
 
 def get_quantizable_ops_recursively(model, example_inputs):  # pragma: no cover
@@ -176,6 +360,9 @@ def get_quantizable_ops_recursively(model, example_inputs):  # pragma: no cover
         cfgs = json.load(f)
         default_cfgs = {}
         fuse_ops = []
+        op_infos_from_cfgs = {}
+        output_tensor_id_op_name = {}
+
         if ipex_ver.release < Version("1.12.0").release:  # pragma: no cover
             default_cfgs = copy.deepcopy(cfgs)
             fuse_ops = get_fuse_ops(cfgs)
@@ -191,6 +378,7 @@ def get_quantizable_ops_recursively(model, example_inputs):  # pragma: no cover
                             break
                     if not re_flag:
                         quantizable_ops.append((op_cfg["id"], op_cfg["name"]))
+
         else:
             (
                 ops_name,
@@ -242,7 +430,7 @@ def get_quantizable_ops_recursively(model, example_inputs):  # pragma: no cover
     logger.info(attention_block)
     logger.info("FFN Blocks : ")
     logger.info(ffn_blocks)
-    return quantizable_ops, cfgs, default_cfgs, fuse_ops
+    return quantizable_ops, cfgs, default_cfgs, fuse_ops, op_infos_from_cfgs, output_tensor_id_op_name
 
 
 def simple_inference(q_model, example_inputs, iterations=1):
@@ -671,3 +859,67 @@ class TransformerBasedModelBlockPatternDetector:  # pragma: no cover
                     if ffn_block:
                         ffn_block_lst.append(ffn_block)
         return attention_block_lst, ffn_block_lst
+
+
+class CpuInfo(object):  # pragma: no cover
+    """Get CPU Info."""
+
+    def __init__(self):
+        """Get whether the cpu numerical format is bf16, the number of sockets, cores and cores per socket."""
+        self._bf16 = False
+        self._vnni = False
+        info = cpuinfo.get_cpu_info()
+        if "arch" in info and "X86" in info["arch"]:
+            cpuid = cpuinfo.CPUID()
+            max_extension_support = cpuid.get_max_extension_support()
+            if max_extension_support >= 7:
+                ecx = cpuid._run_asm(
+                    b"\x31\xC9",  # xor ecx, ecx
+                    b"\xB8\x07\x00\x00\x00" b"\x0f\xa2" b"\x89\xC8" b"\xC3",  # mov eax, 7  # cpuid  # mov ax, cx  # ret
+                )
+                self._vnni = bool(ecx & (1 << 11))
+                eax = cpuid._run_asm(
+                    b"\xB9\x01\x00\x00\x00",  # mov ecx, 1
+                    b"\xB8\x07\x00\x00\x00" b"\x0f\xa2" b"\xC3",  # mov eax, 7  # cpuid  # ret
+                )
+                self._bf16 = bool(eax & (1 << 5))
+        if "arch" in info and "ARM" in info["arch"]:  # pragma: no cover
+            self._sockets = 1
+        else:
+            self._sockets = self.get_number_of_sockets()
+        self._cores = psutil.cpu_count(logical=False)
+        self._cores_per_socket = int(self._cores / self._sockets)
+
+    @property
+    def bf16(self):
+        """Get whether it is bf16."""
+        return self._bf16
+
+    @property
+    def vnni(self):
+        """Get whether it is vnni."""
+        return self._vnni
+
+    @property
+    def cores_per_socket(self):
+        """Get the cores per socket."""
+        return self._cores_per_socket
+
+    def get_number_of_sockets(self) -> int:
+        """Get number of sockets in platform."""
+        cmd = "cat /proc/cpuinfo | grep 'physical id' | sort -u | wc -l"
+        if psutil.WINDOWS:
+            cmd = r'wmic cpu get DeviceID | C:\Windows\System32\find.exe /C "CPU"'
+
+        with subprocess.Popen(
+            args=cmd,
+            shell=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT,
+            universal_newlines=False,
+        ) as proc:
+            proc.wait()
+            if proc.stdout:
+                for line in proc.stdout:
+                    return int(line.decode("utf-8", errors="ignore").strip())
+        return 0
